image_similarity_app/requirements.txt
blinker==1.9.0
click==8.3.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
filelock==3.13.1
Flask==3.1.2
fsspec==2024.6.1
ftfy==6.3.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pillow==11.0.0
regex==2025.9.1
setuptools==70.2.0
sympy==1.13.3
torch==2.8.0+cpu
torchaudio==2.8.0+cpu
torchvision==0.23.0+cpu
tqdm==4.67.1
typing_extensions==4.12.2
wcwidth==0.2.13
Werkzeug==3.1.3
